Remove null or empty checks
