+++ Validator/Validator.csproj
@@ -45,8 +45,6 @@
     <Compile Include="IpVersion.cs" />
     <Compile Include="IsBase64.cs" />
     <Compile Include="IsCreditCard.cs" />
-    <Compile Include="IsDate.cs" />
-    <Compile Include="IsJson.cs" />
     <Compile Include="IsNumeric.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Validator.cs" />
